Dto e Filtri categorizzati per migliorare la ricerca

+++ apiPB/Dto/Request/PrelMatDto/PrelMatRequestDto.cs
@@ -14,6 +14,6 @@ namespace apiPB.Dto.Request
 
         public string BarCode { get; set; } = null!;
 
-        public int? PrelQty { get; set; }
+        public double? PrelQty { get; set; }
     }
 }
